### 2021-10-07 #3807 @whedon
[PRE REVIEW]: The 3D Slicer RVXLiverSegmentation plug-in for interactive liver anatomy reconstruction from medical images
**Submitting author:** @antoinevacavant (<a href="http://orcid.org/0000-0001-9616-3282">Antoine Vacavant</a>)
**Repository:** <a href="https://github.com/R-Vessel-X/SlicerRVXLiverSegmentation" target ="_blank">https://github.com/R-Vessel-X/SlicerRVXLiverSegmentation</a>
**Version:** 1.0.0
