### 2021-11-17 #3807 @fedorov
@Kevin-Mattheus-Moerman, I am assuming you will make a separate [REVIEW] issue for the assigned reviewers, right (like you did for the submission I reviewed in the past [here](https://github.com/openjournals/joss-reviews/issues/2006#issuecomment-640799592))?

I am just not sure if the process is the same as before, considering you assigned @deepakri201 19 days ago, but she has not received any instructions.
