### 2021-11-17 #3807 @Kevin-Mattheus-Moerman
@thewtex @ljod @FMalmberg would you be interested in reviewing this work for JOSS ([the Journal of Open Source Software](https://joss.theoj.org)): _"The 3D Slicer RVXLiverSegmentation plug-in for interactive liver anatomy reconstruction from medical images"_?
The [JOSS review process](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) focuses on [the software](https://github.com/R-Vessel-X/SlicerRVXLiverSegmentation) as well as [a short paper](https://github.com/openjournals/joss-papers/blob/joss.03807/joss.03807/10.21105.joss.03807.pdf).

If you haven't reviewed for JOSS before, below is an example of our GitHub based review process, where you can see that sets of check-boxes guide the reviewers through the process:
https://github.com/openjournals/joss-reviews/issues/2969
