### 2021-11-17 #3920 @whedon
[REVIEW]: The 3D Slicer RVXLiverSegmentation plug-in for interactive liver anatomy reconstruction from medical images
**Submitting author:** @antoinevacavant (<a href="http://orcid.org/0000-0001-9616-3282">Antoine Vacavant</a>)
**Repository:** https://github.com/R-Vessel-X/SlicerRVXLiverSegmentation
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.2
