### 2021-11-17 #3920 @deepakri201
Hi @antoinevacavant, could you please confirm that the installation instructions you have provided are correct? For instance, I had to add the two subfolders (RVXLiverSegmentation and RVXLiverSegmentationEffect) instead of the main directory. Thank you!
