### 2021-11-18 21e9a6f
DOC: Fix path to add to the list of additional modules for manual installation
