### 2021-11-18 #3920 @antoinevacavant
Hi @deepakri201 yes the documentation contains an old instruction about this main directory. Sorry for the confusion.
You should add both paths to RVXLiverSegmentation and RVXLiverSegmentationEffect indeed.
We will update the readme file for making this point clear.
