### 2021-12-23 #3920 @deepakri201
Hi @antoinevacavant, I believe there are a few things to be checked for the paper:

- [ ] A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is? --> I believe this section in the paper should be rewritten, as it does not address the statement of need, and includes information more relevant to the use of the extension.
- [ ] References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax? --> The first reference is incomplete.
