### 2021-12-23 R-Vessel-X/SlicerRVXLiverSegmentation#1 @deepakri201
Documentation -- Functionality documentation improvements
Hi, I am currently working on the JOSS review for your extension and have some concerns about the functionality documentation:

- If possible, I would rewrite each section with an overview, and bullet points or a numbered list describing the steps the user has to take. This would be especially useful for the portal veins segmentation description.
- I think it would also be helpful to include screenshots and videos of the sample dataset that is described in your documentation (I am not sure if the same one is used). For instance, a video detailing the selection of nodes for the portal veins would be useful to a new user.
- There seems to be minor discrepancies, for instance the name of the extension is not consistent in the documentation and in the screenshots.
- For the testing section, it would be useful to include what errors might occur, how to resolve them, and what the user should expect after running Reload and Test.
- It would be helpful to describe the parameters for the portal veins (vesselness filter options)
- In exporting the results, there are some files that are exported but not explained in your documentation.

Thank you!
