### 2021-12-23 R-Vessel-X/SlicerRVXLiverSegmentation#2 @deepakri201
Documentation -- Community guidelines missing
Hi, I believe that the section on community guidelines is missing from your documentation:

Thank you
