### 2022-01-07 9cf9442
DOC: Add contributing documentation
