### 2022-01-07 R-Vessel-X/SlicerRVXLiverSegmentation#2 @Thibault-Pelletier
Hi @deepakri201, thank you for opening this issue!
I have just added our CONTRIBUTING.md file.
Please don't hesitate to let me know if you have any feedback on its content!
