### 2022-01-07 R-Vessel-X/SlicerRVXLiverSegmentation#1 @Thibault-Pelletier
Hi @deepakri201, thank you for this detailed feedback on our documentation.
We will be working on it in the following few weeks and we will keep you updated when everything has been pushed to the repo.
