### 2022-01-07 41d2997
ENH: Add support for MRI liver segmentation
* Add preprocessing for MRI images
* Add option to select volume modality
