+++ RVXLiverSegmentationEffect/RVXLiverSegmentationEffectLib/SegmentEditorEffect.py
@@ -6,7 +6,8 @@ import monai
 from monai.inferers.utils import sliding_window_inference
 from monai.networks.layers import Norm
 from monai.networks.nets.unet import UNet
-from monai.transforms import (AddChanneld, Compose, Orientationd, ScaleIntensityRanged, Spacingd, ToTensord)
+from monai.transforms import (AddChanneld, Compose, Orientationd, ScaleIntensityRanged, Spacingd, ToTensord, Resized,
+                              Resize, CropForegroundd, ScaleIntensityRange)
 from monai.transforms.compose import MapTransform
 from monai.transforms.post.array import AsDiscrete, KeepLargestConnectedComponent
 import numpy as np
@@ -24,7 +25,8 @@ class SegmentEditorEffect(AbstractScriptedSegmentEditorEffect):
 
   def __init__(self, scriptedEffect):
     self.device = qt.QComboBox()
-    scriptedEffect.name = 'Segment CT Liver'
+    self.modality = qt.QComboBox()
+    scriptedEffect.name = 'Segment CT/MRI Liver'
     scriptedEffect.perSegment = True  # this effect operates on a single selected segment
     AbstractScriptedSegmentEditorEffect.__init__(self, scriptedEffect)
     self.logic = SegmentEditorEffectLogic()
@@ -48,7 +50,7 @@ class SegmentEditorEffect(AbstractScriptedSegmentEditorEffect):
     return qt.QIcon()
 
   def helpText(self):
-    return "<html>Segments the liver using a UNet model in CT modality Volumes<br><br>" \
+    return "<html>Segments the liver using a UNet model in CT/MRI modality Volumes<br><br>" \
            "A ROI may be necessary to limit memory consumption.</html>"
 
   def setupOptionsFrame(self):
@@ -59,6 +61,9 @@ class SegmentEditorEffect(AbstractScriptedSegmentEditorEffect):
     # CPU / CUDA options
     self.device.addItems(["cuda", "cpu"])
     self.scriptedEffect.addLabeledOptionsWidget("Device:", self.device)
+
+    self.modality.addItems(["CT","MRI"])
+    self.scriptedEffect.addLabeledOptionsWidget("Modality:",self.modality)
 
     # Add ROI options
     self.roiSelector.nodeTypes = ['vtkMRMLMarkupsROINode']
@@ -98,7 +103,7 @@ class SegmentEditorEffect(AbstractScriptedSegmentEditorEffect):
     slicer.vtkSlicerSegmentationsModuleLogic.CopyOrientedImageDataToVolumeNode(self.getClippedMasterImageData(),
                                                                                masterVolumeNode)
     try:
-      self.logic.launchLiverSegmentation(masterVolumeNode, use_cuda=self.device.currentText == "cuda")
+      self.logic.launchLiverSegmentation(masterVolumeNode, use_cuda=self.device.currentText == "cuda", modality=self.modality.currentText)
 
       self.scriptedEffect.saveStateForUndo()
       self.scriptedEffect.modifySelectedSegmentByLabelmap(
@@ -151,6 +156,20 @@ class SegmentEditorEffect(AbstractScriptedSegmentEditorEffect):
 
     roiNode.SetDisplayVisibility(not roiNode.GetDisplayVisibility())
 
+class Normalized(MapTransform):
+  """
+  Normalizes input
+  """
+  def __init__(self, keys, meta_key_postfix: str = "meta_dict") -> None:
+    super().__init__(keys)
+    self.meta_key_postfix = meta_key_postfix
+    self.keys = keys
+
+  def __call__(self, volume_node):
+    d = dict(volume_node)
+    for key in self.keys:
[…]
