### 2022-02-03 #3920 @antoinevacavant
@Kevin-Mattheus-Moerman thanks for your message, I am editing the .md and .bib files of our paper.
Once they are ready, I suppose I have to:
- push these files on the Github project
- recompile the JOSS article (PDF)
- resubmit the paper or change the submission with the new file?
Thanks for clarifying me the process ;)
