### 2022-02-03 #3920 @Kevin-Mattheus-Moerman
@antoinevacavant Please push all changes to the branch under review. Once done you can recompile the paper here by calling the following in a comment:
`@whedon generate pdf` or `@whedon generate pdf from branch branchName`

Once you do the above you are done, so there is no need to resubmit.
