+++ Paper/paper.md
@@ -40,9 +40,13 @@ bibliography: paper.bib
 
 # Summary
 
-Annotation plays a key role in the creation of reference datasets that are useful to evaluate medical image processing algorithms and to train machine learning based architectures. `RVXLiverSegmentation` is a `3D Slicer` [@Kikinis2014-3DSlicer;@3DSlicer2020-Web] plug-in aiming at speeding-up the annotation of liver anatomy from medical images (CT -Computerized Tomography- scans or MRI -Magnetic Resonance Imaging- for instance). This organ has particular anatomy and physiology; within its parenchymal volume, the liver receives blood from the portal vein and hepatic artery (the former one being the most visible in medical images), and delivers filtered blood through the hepatic veins, toward inferior vena cava. The blood vessels subdivide into the liver as fine vascular tree structures, which make the segmentation difficult, mostly in MRI modality. To facilitate this task, our plug-in is decomposed into modules dedicated to the segmentation of the liver volume, inner vessels and possible tumors. `RVXLiverSegmentation` can be downloaded from [@plugin] or installed from the `3D Slicer` software as an official module. 
+`RVXLiverSegmentation` is a `3D Slicer` [@Kikinis2014-3DSlicer;@3DSlicer2020-Web] plug-in aiming at speeding-up the annotation of liver anatomy from medical images (CT -Computerized Tomography- scans or MRI -Magnetic Resonance Imaging- for instance). This organ has particular anatomy and physiology; within its parenchymal volume, the liver receives blood from the portal vein and hepatic artery (the former one being the most visible in medical images), and delivers filtered blood through the hepatic veins, toward inferior vena cava. The blood vessels subdivide into the liver as fine vascular tree structures, which make the segmentation difficult, mostly in MRI modality. To facilitate this task, our plug-in is decomposed into modules dedicated to the segmentation of the liver volume, inner vessels and possible tumors. `RVXLiverSegmentation` can be downloaded from [@plugin] or installed from the `3D Slicer` software as an official module. 
 
 # Statement of need
+
+Annotation plays a key role in the clinical practice and in the creation of reference datasets that are useful to evaluate medical image processing algorithms and to train machine learning based architectures [@Park2020]. Since softwares embedded into acquisition machines are not always accurate for this task, researchers and engineers have developed more and more interactive annotation and segmentation tools [@Wang2021]. But, despite these developments, annotated datasets are still missing for many clinical applications. The creation of the `RVXLiverSegmentation` plug-in has been motivated by the dramatic absence of a public dataset providing hepatic MRI images together with their ground-truth segmentations of liver anatomy. By creating our own dataset, we have developed a plug-in specificly dedicated to this task, that can be used publicly by any user for the construction of new datasets related to liver pathologies. Moreover, our plug-in has numerous clinical applications, such as the hepatic volumetry, which is becoming increasingly important in view of the growing number of chronic hepatopathies requiring transplants or hepatectomies [@Lim2014]. 
+
+# Overview of RVXLiverSegmentation
 
 For research purpose needing annotation of liver anatomy from medical images, the `RVXLiverSegmentation` provides 7 main tabs:
 
