+++ Paper/paper.bib
@@ -20,11 +20,11 @@ year = {2008},
 }
 
 @misc{3DSlicer2020-Web,
-author = {{---}},
+author = {Kikinis, R. and Pieper, S.D. and Fillion-Robin, J.P.},
 title = {{3D Slicer}},
-year = {2021},
+year = {2022},
 url = {https://www.slicer.org},
-note = {Accessed: 10/01/2021},
+note = {Accessed: 01/02/2022},
 }
 
 @inproceedings{Lamy2020-VPH,
@@ -63,10 +63,39 @@ year = {2021},
 @misc{plugin,
   author = {T. Pelletier and J. Finet and A. Vacavant},
   title = {RVesselX Slicer Liver Anatomy Annotation Plugin},
-  year = {2021},
+  year = {2022},
   publisher = {GitHub},
   journal = {GitHub repository},
   url = {https://github.com/R-Vessel-X/SlicerRVXLiverSegmentation},
-  note = {Accessed: 10/01/2021},  
+  note = {Accessed: 01/02/2022},  
+}
+
+@article{Park2020,
+  author = {Park, S. and Chu, L.C. and Fishman, E.K. and Yuille, A.L. and Vogelstein, B. and Kinzler, K.W. and Horton, K.M. and Hruban, R.H. and Zinreich, E.S. and Fouladi, D.F. and Shayesteh, S. and Graves, J. and Kawamoto, S.},
+  title = {Annotated normal CT data of the abdomen for deep learning: Challenges and strategies for implementation},
+  journal = {Diagnostic and Interventional Imaging},
+  year = {2020},
+  volume = {101},
+  number = {1},
+  pages = {35-44},
+}
+
+@article{Wang2021,
+author = {Wang, Shanshan and Li, Cheng and Wang, Rongpin and Liu, Zaiyi and Wang, Meiyun and Tan, Hongna and Wu, Yaping and Liu, Xinfeng and Sun, Hui and Yang, Rui and Liu, Xin and Chen, Jie and Zhou, Huihui and Ben Ayed, Ismail and Zheng, Hairong},
+title = {Annotation-efficient deep learning for automatic medical image segmentation},
+journal = {Nature Communications},
+volume = {12},
+number = {5915},
+year = {2021},
+}
+
+@article{Lim2014,
+author = {Lim, M.C. and Tan, C.H. and Cai, J. and Zheng, J. and Kow, A.W.},
+title = {{CT} volumetry of the liver: where does it stand in clinical practice?},
+journal = {Clinical Radiology},
+year = {2014},
+volume = {69},
+number = {9},
+pages = {887--95},
 }
 
