### 2022-02-10 #3920 @antoinevacavant
@Kevin-Mattheus-Moerman We have revised/generated the paper, and solved a first issue for the repo. We need more time to solve the second issue (tutorial video, etc.).
Would it be possible to continue the review process despite this unsolved issue?
Thanks.
