### 2022-02-11 #3920 @Kevin-Mattheus-Moerman
@antoinevacavant also note the reviews for JOSS are quite fluid so issues can be resolved simultaneously or incrementally / one at a time, irrespective of what reviewer raised them. So the review remains "ongoing" until all points for all reviewers are addressed, So there are no review "rounds" like for a classic journal. Does this help clarify things?
