### 2022-02-13 #3920 @antoinevacavant
@Kevin-Mattheus-Moerman
This issue is solved: https://github.com/R-Vessel-X/SlicerRVXLiverSegmentation/issues/2 (Community guidelines)
This one is still unresolved: https://github.com/R-Vessel-X/SlicerRVXLiverSegmentation/issues/1
And we addressed the points raised by @deepakri201 in the new version of the paper (statement of need + references).
