### 2022-03-08 #3920 @deepakri201
I left the one box unticked because I believe there still are a few comments to be addressed here: https://github.com/R-Vessel-X/SlicerRVXLiverSegmentation/issues/1
