### 2022-03-24 R-Vessel-X/SlicerRVXLiverSegmentation#4 @lassoan
Update installation instructions
[Installation instructions](https://github.com/R-Vessel-X/SlicerRVXLiverSegmentation#manually-installing-the-plugin) are not up-to-date:
- [ ] Refer to this documentation page: https://slicer.readthedocs.io/en/latest/user_guide/extensions_manager.html#install-extensions instead of the old wiki
- [ ] Describe that the user should search for `RVesselX` extension name
- [ ] Move manual installation instructions (and Testing section) into a `Developers` section at the bottom of the page (or on a separate page)
