### 2022-03-24 #3920 @lassoan
I've reviewed the paper, tested the software, and went through the checklist. All checks passed but I could not tick the items (probably because I did not accept the invitation in time and now the invitation is expired).

I had some difficulty with testing vessel segmentation part because I'm not familiar enough with liver vasculature, but the feature seemed to work. Fixing @deepakri201's comments at https://github.com/R-Vessel-X/SlicerRVXLiverSegmentation/issues/1 will address this issue.

I've added a few suggestions about improving installation instructions: https://github.com/R-Vessel-X/SlicerRVXLiverSegmentation/issues/4
