### 2022-03-24 #3920 @Kevin-Mattheus-Moerman
@lassoan thanks.

We have also since moved onto a different editorial bot system. To fix your box ticking issue, can you call the following in a comment:
`@editorialbot generate my checklist`
