### 2022-03-24 R-Vessel-X/SlicerRVXLiverSegmentation#1 @Kevin-Mattheus-Moerman
@Thibault-Pelletier any updates here?
