### 2022-03-24 #3920 @lassoan
Yes, I would recommend acceptance. Issues https://github.com/R-Vessel-X/SlicerRVXLiverSegmentation/issues/1 and https://github.com/R-Vessel-X/SlicerRVXLiverSegmentation/issues/4 would be nice to be fixed soon but the current state of things is already acceptable for me.
