### 2022-04-14 70f23f6
ENH: Remove obsolete IVC branches and improve doc
* Remove unnecessary IVC branches
* Add image tooltips to help in placing the Portal and IVC nodes
