### 2022-04-14 R-Vessel-X/SlicerRVXLiverSegmentation#4 @Thibault-Pelletier
Hi @lassoan,
Thank you for your review. I have just updated the documentation of the extension. Could you have a go and let me know if you still see anything which needs fixing?

Thanks in advance!
