### 2022-04-14 R-Vessel-X/SlicerRVXLiverSegmentation#1 @Thibault-Pelletier
Hi @Kevin-Mattheus-Moerman and @deepakri201,
Thank you for your feedback and sorry for the delay.
I have just updated the documentation to add more detailed information on using the extension.
We are also in the process of editing the tutorial videos.
They should be made available (hopefully) sometime next week.
