### 2022-04-15 #3920 @antoinevacavant
Hi @Kevin-Mattheus-Moerman thanks a lot for this good news ;)
- Missing DOI's are OK
- I have corrected the paper according to your suggestions, and pushed to the repo
- Version tag 1.0.0 can be kept
- Please note that the documentation is currently updated according to reviews, and a tutorial video will be added soon!
