### 2022-04-15 #3920 @antoinevacavant
@Kevin-Mattheus-Moerman Well I forgot to update correctly the .bib file ;( And last typo is corrected
