+++ Paper/paper.md
@@ -40,7 +40,7 @@ bibliography: paper.bib
 
 # Summary
 
-`RVXLiverSegmentation` is a `3D Slicer` [@Kikinis2014-3DSlicer;@3DSlicer2020-Web] plug-in aiming at speeding-up the annotation of liver anatomy from medical images (CT -Computerized Tomography- scans or MRI -Magnetic Resonance Imaging- for instance). This organ has particular anatomy and physiology; within its parenchymal volume, the liver receives blood from the portal vein and hepatic artery (the former one being the most visible in medical images), and delivers filtered blood through the hepatic veins, toward inferior vena cava. The blood vessels subdivide into the liver as fine vascular tree structures, which make the segmentation difficult, especially in the case of MRI. To facilitate this task, our plug-in is decomposed into modules dedicated to the segmentation of the liver volume, inner vessels and possible tumors. `RVXLiverSegmentation` can be downloaded from [@plugin] or installed from the `3D Slicer` software as an official module. 
+`RVXLiverSegmentation` is a `3D Slicer` [@Kikinis2014-3DSlicer;@3DSlicer2020-Web] plug-in aiming at speeding-up the annotation of liver anatomy from medical images (CT -Computerized Tomography- scans or MRI -Magnetic Resonance Imaging- for instance). This organ has particular anatomy and physiology; within its parenchymal volume, the liver receives blood from the portal vein and hepatic artery (the former one being the most visible in medical images), and delivers filtered blood through the hepatic veins, toward inferior vena cava. The blood vessels subdivide into the liver as fine vascular tree structures, which makes the segmentation difficult, especially in the case of MRI. To facilitate this task, our plug-in is decomposed into modules dedicated to the segmentation of the liver volume, inner vessels and possible tumors. `RVXLiverSegmentation` can be downloaded from [@plugin] or installed from the `3D Slicer` software as an official module. 
 
 # Statement of need
 
