+++ Paper/paper.bib
@@ -7,6 +7,7 @@ editor = {Jolesz, F.A.},
 volume = {3(19)},
 pages = {277--289},
 isbn = {978-1-4614-7656-6},
+doi = {10.1007/978-1-4614-7657-3_19},
 }
 
 @article{a2008-VMTK,
@@ -17,6 +18,7 @@ OPTjournal = {MBEC},
 volume = {46},
 pages = {1097--1112}, 
 year = {2008},
+doi = {10.1007/s11517-008-0420-1},
 }
 
 @misc{3DSlicer2020-Web,
@@ -41,6 +43,7 @@ title = {A Time-dependent joint segmentation and registration model: Application
 booktitle = {IEEE IPTA 2020}, 
 address = {Paris, France}, 
 year = {2020},
+doi = {10.1109/ipta50016.2020.9286658},
 }
 
 @article{Affane2021-MDPI,
@@ -50,6 +53,7 @@ journal = {MDPI Journal of Applied Sciences},
 volume = {11},
 number = {11:4895}, 
 year = {2021},
+doi= {10.3390/app11114895},
 }
 
 @inproceedings{Lamy2020-ICPR,
@@ -58,6 +62,7 @@ title = {Vesselness filters: A survey with benchmarks applied to liver imaging},
 booktitle = {IEEE ICPR 2020}, 
 address = {Milano, Italy}, 
 year = {2021},
+doi = {10.1109/icpr48806.2021.9412362},
 }
 
 @misc{plugin,
@@ -78,6 +83,7 @@ year = {2021},
   volume = {101},
   number = {1},
   pages = {35-44},
+  doi = {10.1016/j.diii.2019.05.008},
 }
 
 @article{Wang2021,
@@ -87,6 +93,7 @@ journal = {Nature Communications},
 volume = {12},
 number = {5915},
 year = {2021},
+doi = {10.1038/s41467-021-26216-9},
 }
 
 @article{Lim2014,
@@ -97,5 +104,6 @@ year = {2014},
 volume = {69},
 number = {9},
 pages = {887--95},
+doi = {10.1016/j.crad.2013.12.021},
 }
 
