### 2022-05-03 #3920 @Kevin-Mattheus-Moerman
@antoinevacavant I've incremented the version tag to match that of the ZENODO archived version. Can you confirm this is correct?
