### 2022-05-06 #3920 @Kevin-Mattheus-Moerman
@antoinevacavant congratulations on getting this work published in JOSS! :tada:

A special thanks to @lassoan and @deepakri201 for their great review efforts! :bow:
